Instructions
Write a Python program that will display a message to ask the user to input the type of residency. The program then reads the residency (‘I’ for in-state, ‘O’ for out-of-state, and ‘G’ for graduate). Based on the residency, the program will use the respective current tuition and respective word of either UNDERGRADUATE or GRADUATE ... this needs an if ... elif ... else ....

The program will then enter a loop and calculate the new tuition for every year for five years and inside the loop display a line of the table that will look as below, formatting output amounts to 2 decimal places for amounts.

Your program should have only one print( ) statement to produce all the five non-heading lines. This is the whole purpose of a loop. Also, the word UNDERGRADUATE in the headings should be replaced by GRADUATE when a ‘G’ is selected.